% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFAN-internal.R
\name{lf_modes}
\alias{lf_modes}
\title{Counting peaks detected after restructuring}
\usage{
lf_modes(x, N = 7)
}
\arguments{
\item{x}{A 'len_freq' object created with lf_create.}

\item{N}{The maximum MA tested is 2N-1.}
}
\description{
Counting peaks detected after restructuring
}
